Renamed main exports to index.tsx to follow front end convention
