+++ client/src/routes/Search.tsx
@@ -1,5 +1,5 @@
-import HelmetWrapper from "components/Helmet/HelmetWrapper";
-import RenderComponentWithSideBar from "components/WithSideBar/WithSideBar";
+import HelmetWrapper from "components/Helmet";
+import RenderComponentWithSideBar from "components/WithSideBar";
 
 function SearchComponent() {
   return <p>Search through your past lessons on Notum</p>
