+++ client/src/routes/Settings.tsx
@@ -1,5 +1,5 @@
-import HelmetWrapper from "components/Helmet/HelmetWrapper";
-import RenderComponentWithSideBar from "components/WithSideBar/WithSideBar";
+import HelmetWrapper from "components/Helmet";
+import RenderComponentWithSideBar from "components/WithSideBar";
 
 function SettingsComponent() {
   return <p>Create your courses on Notum</p>
